requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# examples/app.py: 5
Pillow >= 5.1.0

# examples/app.py: 10
# examples/glfwBackend.py: 3
# fieldanimation/shader.py: 1,2
# fieldanimation/texture.py: 1
PyOpenGL >= 3.1.0

# examples/app.py: 18
# examples/glfwBackend.py: 2
glfw >= 1.6.0

# examples/app.py: 12,13
imgui >= 0.1.0

# examples/app.py: 4
# fieldanimation/shader.py: 4
numpy >= 1.14.3

# setup.py: 7,8
pip >= 9.0.1

# setup.py: 5
setuptools >= 39.0.1
